Measuring nanoTime of rect
Was able to make it so the program gave the average after a certain amount of trials (6 in this case) and the time showed up. However, it seems that something is not quite right with how the time is measured: from the moment the rectangle appears on screen to the moment the spacebar is pressed.